fix addapt array tests to the change

+++ tests/epyccel/modules/arrays.py
@@ -514,7 +514,7 @@ def array_int32_in_bool_out_2d_F_complex_3d_expr( x, y, ri ):
 def array_real_1d_sum_stack_array():
     from numpy import zeros
     a = zeros(10)
-    s = 0.
+    s = np.float64(0.0)
     for i in range(10):
         s += a[i]
     return s
@@ -523,7 +523,7 @@ def array_real_1d_sum_stack_array():
 def array_real_1d_div_stack_array():
     from numpy import ones
     a = ones(10)
-    s = 0.
+    s = np.float64(0.0)
     for i in range(10):
         s += 1.0 / a[i]
     return s
@@ -559,7 +559,7 @@ def multiple_stack_array_2():
 def array_real_2d_sum_stack_array():
     from numpy import zeros
     a = zeros((10, 10))
-    s = 0.
+    s = np.float64(0.)
     for i in range(10):
         for j in range(10):
             s += a[i][j]
@@ -568,8 +568,10 @@ def array_real_2d_sum_stack_array():
 @stack_array('a')
 def array_real_2d_div_stack_array():
     from numpy import full
+    import numpy as np
+
     a = full((10, 10), 2)
-    s = 1.
+    s = np.float64(1.)
     for i in range(10):
         for j in range(10):
             s /= a[i][j]
@@ -581,7 +583,7 @@ def multiple_2d_stack_array_1():
     from numpy import ones, array
     a = ones((2, 5))
     b = array([[1, 3, 5, 7, 9], [11, 13, 17, 19, 23]])
-    s = 0.0
+    s = np.float64(0.0)
     j = 0
     for i in range(2):
         for j in range(5):
@@ -595,7 +597,7 @@ def multiple_2d_stack_array_2():
     a = ones(5)
     b = array([[2, 4, 6, 8, 10], [1, 3, 5, 7, 9]])
     c = array([[1, 3, 5, 7, 9], [2, 4, 6, 8, 10]])
-    s = 0.0
+    s = np.float64(0.0)
     for i in range(2):
         for j in range(5):
             s = s + b[i][j] - a[j] / c[i][j]
